Added collection name validation (it has to be unique)

+++ MyHoard/MyHoard/ViewModels/AddCollectionViewModel.cs
@@ -8,19 +8,23 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace MyHoard.ViewModels
 {
-    public class AddCollectionViewModel : ViewModelBase
+    public class AddCollectionViewModel : ViewModelBase, IHandle<CollectionServiceErrorMessage>
     {
         private string pageTitle;
         private Collection currentCollection;
         private ObservableCollection<string> thumbnails;
         private bool canSave;
+        private readonly IEventAggregator eventAggregator;
 
-
-        public AddCollectionViewModel(INavigationService navigationService, CollectionService collectionService) : base(navigationService, collectionService)
+        public AddCollectionViewModel(INavigationService navigationService, CollectionService collectionService, IEventAggregator eventAggregator)
+            : base(navigationService, collectionService)
         {
+            this.eventAggregator = eventAggregator;
+            eventAggregator.Subscribe(this);
             PageTitle = AppResources.AddCollection;
             Thumbnails = new BindableCollection<string> { "","\uE114", "\uE104", "\uE107", "\uE10F", "\uE113", "\uE116", "\uE119", "\uE128", "\uE13D", "\uE15D", "\uE15E" };
             CurrentCollection = new Collection();
@@ -73,12 +77,18 @@ namespace MyHoard.ViewModels
 
         public void Save()
         {
-            CollectionService.AddCollection(CurrentCollection);
-            NavigationService.UriFor<CollectionListViewModel>().Navigate();
-            while (NavigationService.BackStack.Any())
+            if (CollectionService.AddCollection(CurrentCollection).Id > 0)
             {
-                this.NavigationService.RemoveBackEntry();
+                NavigationService.UriFor<CollectionListViewModel>().Navigate();
+                while (NavigationService.BackStack.Any())
+                {
+                    this.NavigationService.RemoveBackEntry();
+                }
             }
+        }
+        public void Handle(CollectionServiceErrorMessage message)
+        {
+            MessageBox.Show(message.Content);
         }
     }
 }
